requirements.txt
google-generativeai>=0.8.0
tiktoken>=0.5.0
openai>=1.0.0
pandas>=1.5.0
matplotlib>=3.5.0
openpyxl>=3.0.0 